Drop implicit Word to Long conversion

+++ javalib/src/main/scala/java/lang/UnixProcess.scala
@@ -69,8 +69,9 @@ private[lang] class UnixProcess private (
         val ts = stackalloc[timespec]
         val tv = stackalloc[timeval]
         throwOnError(gettimeofday(tv, null), "Failed to set time of day.")
-        val nsec = unit.toNanos(timeout) + TimeUnit.MICROSECONDS.toNanos(tv._2)
-        val sec  = TimeUnit.NANOSECONDS.toSeconds(nsec)
+        val nsec = unit.toNanos(timeout) + TimeUnit.MICROSECONDS.toNanos(
+          tv._2.toLong)
+        val sec = TimeUnit.NANOSECONDS.toSeconds(nsec)
         ts._1 = tv._1 + sec.toWord
         ts._2 =
           (if (sec > 0) nsec - TimeUnit.SECONDS.toNanos(sec) else nsec).toWord
